% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\alias{as_numeric}
\alias{cut.bin}
\alias{cut.character}
\alias{cut.factor}
\title{Bin series}
\usage{
bin(
  x,
  breaks = NULL,
  max = NULL,
  levels = NULL,
  right = TRUE,
  xlast = NULL,
  wlast = NULL,
  xfirst = NULL
)

as_numeric(x, pos = 0, xfirst = NULL, xlast = NULL, wlast = NULL)

\method{cut}{bin}(x, ..., breaks = NULL)

\method{cut}{character}(x, ..., breaks = NULL)

\method{cut}{factor}(x, ..., breaks = NULL)
}
\arguments{
\item{x}{a character or a factor: the first and last characters
should be any of \code{[}, \code{(}, \verb{]}, \verb{)} and the other characters
should be interpreted as two numerical values separated by a
comma,}

\item{breaks}{a numerical vector of breaks which should be a subset
of the initial set of breaks. If only one break is provided,
all the bins with greater values are merged,}

\item{max}{an integer, to create a >= J bin for integer series}

\item{levels}{a character or a list to define, rename or collapse
the levels for a categorical series}

\item{right}{for numerical series, should the interval be closed on
the right ?}

\item{wlast}{in the case where the upper bond is infinite and
\code{xlast} is not provided, the width of the last class is set to
the one of the before last class. If \code{wlast} is provided, it is
set to the width of the before last class times \code{wlast},}

\item{xfirst, xlast}{the center of the first (last) class, if one
wants to specify something different from the average of the
lower and the upper bonds,}

\item{pos}{a numeric between 0 and 1, 0 for the lower bond, 1 for
the upper bond, 0.5 for the center of the class (or any other
value between 0 and 1), which indicates to \code{as_numeric} how the
bins should be coerced to numerical values,}

\item{...}{further arguments}
}
\value{
a series of class \code{bin} with a `type attribute
}
\description{
A new class called \code{bin} is provided ; a \code{bin} is a series that
contain a relatively few distinct values, that can be used to
compute a frequency table. The input series can be either :
\itemize{
\item an \code{integer} series,
\item a \code{contbin} series (a series that contain values as \verb{[20, 40)},
\item a categorical series,
\item a continuous numerical series.
}

The \code{as_numeric} function is relevant for numerical bins. It
converts a numerical bin (an interval) to a value of the underlying
variable defined by its relative position (from 0 lower bound to 1
upper bound in the bin). Specific arguments are provided for the
first and the last bins
}
\details{
The \code{bin} function takes a series as input and returns a series of
class \code{bin}. It has a \code{type} attribute that can be either equal to
\code{integer}, \code{contbin} or \code{categorical}, and performs the following
tasks:
\itemize{
\item for categorical series, the levels can be defined, renamed and
collapsed using the \code{level} argument, - for integer series, a >= J
series can be computed by setting the \code{max} argument to \code{J}, so
that all the values greater or equal than J are collapsed and the
J, J+1, J+2, ... values are replaced by the mean of the values
greater or equal than J, - numerical bin series are kept as is, or
the number of classes can be reduced using the \code{breaks} argument, -
continuous numerical series are transfomed in numerical bin series,
using the \code{breaks} argument (if \code{NULL}, the breaks are
automatically computed) and the \code{right} argument can be set to
\code{TRUE} (the intervals are closed on the right, the default), or to
\code{FALSE} (the intervals are closed on the left).
}
}
\examples{
# get a few values of the `size` series of the `wages` data set
z <- head(wages$size, 10)
bin(z)
# reduce the number of bins
bin(z, breaks = c(20, 50, 100))
# set the right value of the last interval to 50
bin(z, breaks = 50)
# the `children` series of the `rgp` data set
z <- rgp$children
bin(z) |> head(20)
# set a >= 3 cathegory
bin(z, max = 3) |> head(20)
# the `sector` series of the `wages` data set contains a factor
# with levels `industry`, `building`, `business`, `services` and
# `administration`
z <- head(wages$sector)
bin(z)
# change the order of the levels:
bin(z, levels = c("business", "services", "administration", "building", "industry"))
# rename some levels
bin(z, levels = c("business", "services", government = "administration",
                  construction = "building", "industry"))
# collapse some levels
bin(z, levels = list(blue = c("building", "industry"),
                     white = c("business", "services"),
                     government = "administration"))
z <- head(wages$size, 10)
# coerce to a numeric using the center of the bins
as_numeric(z, pos = 0.5)
# special values for the center of the first and of the last bin
as_numeric(z, pos = 0.5, xfirst = 5, xlast = 400)
# same, but indicating that the width of the last class should be
# twice the one of the before last
as_numeric(z, pos = 0.5, xfirst = 5, wlast = 2)
}
\author{
Yves Croissant
}
\keyword{classes}
