% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_table.R
\name{cont_table}
\alias{cont_table}
\alias{joint}
\alias{conditional}
\alias{marginal}
\alias{var_decomp}
\alias{summary.var_decomp}
\alias{regline}
\title{Contingency table}
\usage{
cont_table(
  data,
  y1,
  y2,
  weights = NULL,
  freq = NULL,
  total = FALSE,
  xfirst1 = NULL,
  xlast1 = NULL,
  wlast1 = NULL,
  xfirst2 = NULL,
  xlast2 = NULL,
  wlast2 = NULL
)

joint(x)

conditional(x, y = NULL)

marginal(x, y = NULL, cols = "f", vals = NULL)

var_decomp(x, y)

\method{summary}{var_decomp}(object, ...)

regline(formula, data)
}
\arguments{
\item{data}{a tibble,}

\item{y1, y2}{the two series used the construct the contingency
table, the distinct values of the first and the second will
respectively be the rows and the columns of the contingency
table,}

\item{weights}{a series containing the weights that should be used
to mimic the population,}

\item{freq}{the frequencies (in case where data is a contingency
table),}

\item{total}{if \code{TRUE}, a total is added to the table,}

\item{xfirst1, xfirst2}{the center of the first class for the two
series,}

\item{xlast1, xlast2, }{the center of the last class for the two
series,}

\item{wlast1, wlast2}{the width of the last class for the two
series,}

\item{x, object}{a tibble containing the contingency table}

\item{y}{the series on which the operation should be computed}

\item{cols}{see \code{\link[=freq_table]{freq_table()}},}

\item{vals}{see \code{\link[=freq_table]{freq_table()}},}

\item{...}{further arguments.}

\item{formula}{a formula which describe the model to estimate with
the \code{regline} function,}
}
\value{
a tibble
}
\description{
A contingency table returns the counts of all the combinations of
the modalities of two series in a table for which every modality of
the first series is a row and every modality of the second series
is a column. The \code{joint}, \code{marginal} and \code{conditional} functions
compute these three distributions from the contingency table (by
indicating one series for the last two). \code{mean}, \code{variance},
\code{stdev}, \code{covariance} and \code{correlation} methods are
provided. \code{regline} computes the regression line and \code{var_decomp}
the variance decomposition.
}
\examples{
library("dplyr")
cont_table(employment, education, sex)
cont_table(employment, education, sex, weights = weights)
cont_table(employment, education, sex) \%>\% conditional(sex)
cont_table(wages, wage, size)
cont_table(wages, wage, size) \%>\% joint
cont_table(wages, wage, size) \%>\% joint \%>\% mean
cont_table(wages, wage, size) \%>\% marginal(size)
cont_table(wages, wage, size) \%>\% conditional(size)

}
\author{
Yves Croissant
}
