% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descstat.R
\docType{package}
\name{descstat-package}
\alias{descstat}
\alias{descstat-package}
\title{descstat: a toolbox for descriptive statistics}
\description{
Descriptive statistics consist on presenting the distribution of
series for a sample in tables (frequency table for one series,
contingency tables for two series), ploting this distribution and
computing some statistics that summarise it. \strong{descstat} provides
a complete toolbox to perform this tasks. It has been writen using
the tidyverse conventions, especially the selection of series using
their unquoted names and the use of the pipe operator.
}
\section{The bin class}{


The \code{bin} function takes a series as input that can either:
\itemize{
\item a continouis variable,
\item a categorical series,
\item an integer series,
\item a continuous bin series
}

and returns an object of class \code{bin}. The \code{bin} function is
particularly usefull for continuous series. It enables:
\itemize{
\item creating bins from numerical values, which is performed by the
\code{base::cut} function which turns a numerical series to a bin,
\item coercing bins to numerical values, eg getting from the \verb{[10,20)}
bin the lower bound (10), the upper bound (20), the center (15) or
whatever other value of the bin,
\item reducing the number of bins by merging some of them (for example
\verb{[0,10)}, \verb{[10, 20)}, \verb{[20,30)}, \verb{[30,Inf)} to \verb{[0,20)}, \verb{[20,Inf)}
}

The last task is performed using the \code{as_numeric}
function. Coercing bins to their center values is the basis of the
computation of descripting statistics for bins.
}

\section{Frequency and contingency tables}{


The \code{freq_table} and \code{cont_table} are based on the \code{dplyr::count}
function but offer a much richer interface and performs easily
usual operations which are tedious to obtain with \code{dplyr::count} or
\code{base::table} functions. This includes:
\itemize{
\item adding a total,
\item for frequency tables, computing other kind of frequencies than
the counts, for example relative frequencies, percentage,
cummulative frequencies, etc.,
\item for contingency tables, computing easily the joint, marginal and
conditional distributions,
\item printing easily the contingency table as a double entry table.
}
}

\section{Plotting the distribution}{


\strong{descstat} uses the \strong{tinyplot} package to provide classic plots
for univariate or bivariate distributions. This includes histogram,
frequency plot, pie chart, cummulative plot and Lorenz curve. More
precisely, \code{type_dsc_###} functions are provided and the relevant
plot is obtained using \verb{tinyplot(..., type = type_dsc_###)}.
}

\section{Descriptive statistics}{


A full set of statistical functions (of central tendency,
dispersion, shape, concentration and covariation) are provided and
can be applied directly on objects of class \code{freq_table} or
\code{cont_table}. Some of them are methods of generics defined by the
\code{base} or \code{stats} package, some other are defined as methods for
generics function provided by the \strong{descstat} function when the
corresponding \strong{R} function is not generic. For example,
\itemize{
\item \code{mean} is generic, so that we wrote a
\code{mean.freq_table} method to compute directly the mean of a series
from a frequency table.
\item \code{var} is not generic, so that we provide the \code{variance} generic
and a method for \code{freq_table} objects.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://www.r-project.org}
}

}
\author{
\strong{Maintainer}: Yves Croissant \email{yves.croissant@univ-reunion.fr}

}
\keyword{package}
