% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deststat.R
\name{descstat}
\alias{descstat}
\title{Speedrun Descriptive Statistics}
\usage{
descstat(data, group, numeric_variable, na.rm = TRUE)
}
\arguments{
\item{data}{the dataset you would like to examine}

\item{group}{the column to group by}

\item{numeric_variable}{the numeric column to summarize}
}
\value{
descriptive statistics in a tibble with one row per group and summary statistics in columns
}
\description{
This function produces a table that contains descriptive statistics (min, max, mean, median, and sd) of a numeric variable grouped by any column with repeated values. This analysis will remove any missing values from the data.
}
\examples{
descstat(gapminder, year, lifeExp)
descstat(cancer_sample, diagnosis, area_mean)

}
