% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{employment}
\alias{employment}
\title{French employment survey}
\format{
a tibble containing:
\itemize{
\item activity : a factor with levels \code{occupied}, \code{unemployed} and \code{inactive}
\item time : job time a factor with levels \code{part}, \code{full} and
\code{unknown}
\item education : level of education
\item age : age in years
\item sex : one of \code{male} or \code{female}
\item household : kind of household, \code{single}, \code{monop} (monoparental
family), \code{couple} (couple without children), \code{family} (couple with
families) and \code{other}.
\item weights : weights to mimic the population
}
}
\source{
Employment survey 2018, site of the \href{https://www.insee.fr/fr/statistiques/4191029}{INSEE}
}
\description{
The employment survey gives information about characteristics of
individuals (employed/unemployed, part/full time job, education,
etc.)
}
