% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{generics}
\alias{generics}
\alias{variance}
\alias{gmean}
\alias{covariance}
\alias{correlation}
\alias{stdev}
\alias{madev}
\alias{modval}
\alias{medial}
\alias{kurtosis}
\alias{skewness}
\alias{variance.default}
\alias{gmean.default}
\alias{stdev.default}
\alias{skewness.default}
\alias{kurtosis.default}
\alias{pre_print}
\title{Generic functions provided by the statecoaes package}
\usage{
variance(x, ...)

gmean(x, r = 1, ...)

covariance(x, ...)

correlation(x, ...)

stdev(x, ...)

madev(x, ...)

modval(x, ...)

medial(x, ...)

kurtosis(x, ...)

skewness(x, ...)

\method{variance}{default}(x, w = NULL, ...)

\method{gmean}{default}(x, r = 1, ...)

\method{stdev}{default}(x, w = NULL, ...)

\method{skewness}{default}(x, ...)

\method{kurtosis}{default}(x, ...)

pre_print(x)
}
\arguments{
\item{x}{the main argument, basically a character containing the
classes}

\item{...}{further arguments}

\item{r}{the order of the mean for the \code{gmean} function}

\item{w}{a vector of weights}

\item{center}{the center value used to compute the deviations, one
of \code{"median"} or \code{"mean"}}
}
\value{
a numeric or a tibble
}
\description{
Some generics were created, some of them don't exist in base R (to
compute the mode and the mediale of a distribution for example) or
because the functions exist but are not generic (this is the case
of \code{var}, \code{sd}, \code{mad}, \code{cov} and \code{cor} which are replaced by our
\code{variance}, \code{stdev}, \code{madev}, \code{covariance} and \code{correlation}
generic functions
}
\author{
Yves Croissant
}
