% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_table.R
\name{hist_table.methods}
\alias{hist_table.methods}
\alias{mean.hist_table}
\alias{variance.hist_table}
\alias{stdev.hist_table}
\alias{madev.hist_table}
\alias{modval.hist_table}
\alias{quantile.hist_table}
\alias{median.hist_table}
\alias{medial.hist_table}
\alias{gini}
\title{Methods for hist_table objects}
\usage{
\method{mean}{hist_table}(x, ...)

\method{variance}{hist_table}(x, ...)

\method{stdev}{hist_table}(x, ...)

\method{madev}{hist_table}(x, center = c("median", "mean"), ...)

\method{modval}{hist_table}(x, ...)

\method{quantile}{hist_table}(x, y = c("value", "mass"), probs = c(0.25, 0.5, 0.75), ...)

\method{median}{hist_table}(x, ..., y = c("value", "mass"))

\method{medial}{hist_table}(x, ...)

gini(x)
}
\arguments{
\item{x}{a hist_table object,}

\item{...}{further arguments}

\item{center}{for the \code{madev} method, this can equal \code{"median"}
(the default) or \code{"mean"}}

\item{y}{for the quantile method, one of \code{"value"} or \code{"mass"}}

\item{probs}{the probabilities for which the quantiles have to be
computed}
}
\value{
a  numeric
}
\description{
Functions and methods to compute the median, the mean, the mode,
the medial and quantiles for hist_table objects
}
\examples{

z <- Salaires \%>\% hist_table(salaire)
z \%>\% median
z \%>\% medial
z \%>\% modval
z \%>\% quantile(probs = c(0.25, 0.5, 0.75))
z \%>\% quantile(y = "mass", probs = c(0.25, 0.5, 0.75))
}
\author{
Yves Croissant
}
