% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{income}
\alias{income}
\title{Income of French households}
\format{
a tibble containing :
\itemize{
\item bin: bin of income,
\item number: number of households in the bin,
\item income: mass of income in the bin.
}
}
\source{
Impot sur le revenu par commune (IRCOM)
\href{https://www.impots.gouv.fr/portail/statistiques}{DGI}'s
website.
}
\description{
Bins of income classes, number of households and mass of income.
}
\keyword{datasets}
