% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{indexes}
\alias{indexes}
\title{Composite indexes}
\usage{
indexes(data, an, bien, quant, prix, base, chained = FALSE)
}
\arguments{
\item{data}{a tibble}

\item{an}{the date of observation}

\item{bien}{the series containing the goods that are part of the
computation of the indexes}

\item{quant}{the quantity}

\item{prix}{the price}

\item{base}{the base date}

\item{chained}{if \code{TRUE}, chained indexes are computed}
}
\value{
a tibble
}
\description{
Computation of composite indexes (Laspeyres, Paasche and Fisher),
chained or not.
}
\author{
Yves Croissant
}
