% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_plot.R
\name{pre_plot}
\alias{pre_plot}
\alias{pre_plot.freq_table}
\alias{pre_plot.cont_table}
\title{Put a tibble in form to plot}
\usage{
pre_plot(x, y = NULL, plot = NULL, ...)

\method{pre_plot}{freq_table}(
  x,
  y = NULL,
  plot = c("histogram", "freqpoly", "lorenz", "stacked", "cumulative"),
  ...
)

\method{pre_plot}{cont_table}(x, ...)
}
\arguments{
\item{x}{a tibble returned by the \code{freq_table} or the \code{cont_table}
function, which should contain the center of the classes (\code{x})
and at least one measure of the frequencies or densities (one
of \code{f}, \code{n}, \code{p}, \code{d}),}

\item{y}{mandatory argument if the tibble contains more than one
frequency or density,}

\item{plot}{for object of class \code{freq_table} one of \code{histogram} (a
tibble is returned with columns \code{x}, \code{y}, \code{xend}, \code{yend} that
should be passed to \code{geom_polygon}), \code{freqpoly} (a tibble with
columns \code{x} and \code{y} that should be passed to \code{geom_line}),
\code{stacked} (to compute a stacked bar plot or a pie chart usign
\code{geom_col}), \code{cumulative} (which returns a tibble with \code{x},
\code{y}, \code{xend} and \code{yend} that should be passed to \code{geom_segment})
and \code{lorenz} for the Lorenz curve,}

\item{...}{further arguments.}
}
\value{
a tibble
}
\description{
Convert a tibble built using \code{freq_table}, or \code{cont_table} in a
shape that makes it easy to plot.
}
\examples{
library("dplyr")
library("ggplot2")
pad <- padova \%>\%
       freq_table(price, breaks = c(100, 200, 300, 400, 500, 1000),
       right = TRUE, cols = "Npd")
pad \%>\% pre_plot(y = "d") \%>\% ggplot() + geom_polygon(aes(x, y))
pad \%>\% pre_plot(y = "d", plot = "freqpoly") \%>\%
ggplot() + geom_line(aes(x, y))
## A pie chart
wages \%>\% freq_table(sector, "p", total = FALSE) \%>\%
  pre_plot("p", plot = "stacked") \%>\% ggplot(aes(x = 2, y = p, fill = sector)) +
  geom_col() + geom_text(aes(y = ypos, label = round(p))) +
  coord_polar(theta = "y")

}
\author{
Yves Croissant
}
