% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rgp}
\alias{rgp}
\title{Extract of the French census}
\format{
a tibble containing :
\itemize{
\item cars : number of cars
\item rooms : number of rooms of the housing
\item children : number of children
\item type : type of household ; \code{couple} or \code{monop} (for mono-parental families)
\item ER : sex if the R^th^ child (from 1 to 11).
}
}
\source{
\href{https://www.insee.fr/fr/statistiques/4507685?sommaire=4508161}{site of the INSEE}
}
\description{
this extract gives information about a sample of French households
}
