% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyplot.R
\name{tinyplot}
\alias{tinyplot}
\alias{type_pie}
\alias{type_lorenz}
\alias{type_freqpoly}
\alias{type_histo}
\alias{type_cumul}
\alias{type_cont.table}
\alias{type_anova}
\title{tinyplot types}
\usage{
type_pie(
  position = NULL,
  edges = 200,
  hole = 0,
  radius = 1,
  init.angle = 0,
  f = c("none", "n", "f", "p"),
  digits = 2,
  pal = "Set2"
)

type_lorenz()

type_freqpoly(breaks = NULL)

type_histo(wlast = 2, breaks = NULL)

type_cumul(expand = 0.1, geom = NULL, breaks = NULL, max = NULL)

type_cont.table(size = c(1, 2))

type_anova(level = 0.95, length = 0.1, expand = 0.05)
}
\arguments{
\item{position}{the position of the labels for \code{type_pie}}

\item{edges}{the number of edges for \code{type_pie}}

\item{hole}{the size of the hole of a donut chart for \code{type_pie}}

\item{radius}{the size of the foreground for \code{type_pie}}

\item{init.angle}{the initital angle for \code{type_pie}}

\item{f}{the frequency that is printed, one of \code{"none"} (the
default), \code{"n"} for the absolute frequency, \code{"f"} for the
relative frequency and \code{"p"} for the percentages}

\item{digits}{the number of printed digits of the frequencies}

\item{pal}{a palette of colors}

\item{wlast}{comment}

\item{expand}{for \code{type_cumul} a numeric indicating the fraction of
the range of the series and set the width of the segments for
which the cummulative distribution is 0 or 1, for \code{type_anova},
a numeric of length 2 indicating the horizontal and vertical
margins added to the range of the series}

\item{geom}{for \code{type_cumul} the geometry that is ploted ; by
default, \code{"s"} for integer series and \code{"b"} for continuous bin
series}

\item{max, breaks}{see \code{descstat::freq_table}}

\item{size}{the sizes of the smallest and of the largest point for
\code{type_cont.table}}

\item{level}{the confidence interval for \code{type_anova}}

\item{length}{the length of the error bars for \code{type_anova}}
}
\value{
invisible, used for its side effects
}
\description{
tinyplot types for frequency tables
}
\author{
Yves Croissant
}
