% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls.R
\name{cls2val}
\alias{cls2val}
\alias{cls2val.character}
\alias{cls2val.factor}
\alias{cls2val.cont_table}
\title{Convert class to values}
\usage{
cls2val(x, pos = 0, xfirst = NULL, xlast = NULL, inflate = NULL, ...)

\method{cls2val}{character}(x, pos = 0, xfirst = NULL, xlast = NULL, inflate = NULL, ...)

\method{cls2val}{factor}(x, pos = 0, xfirst = NULL, xlast = NULL, inflate = NULL, ...)

\method{cls2val}{cont_table}(x, y = 1, pos = 0.5, ...)
}
\arguments{
\item{x}{a series that contains a class of values, the first and
last characters should be any of \code{[}, \code{(}, \verb{]}, \verb{)} and the
other characters should be interpreted as two numerical values
separated by a \verb{,}}

\item{pos}{a numeric between 0 and 1, 0 for the lower bond, 1
for the upper bond, 0.5 for the center of the class (and any
other value between 0 and 1)}

\item{xfirst}{center of the first class, if one wants to specifie
something different from the average of the lower and the upper
bonds}

\item{xlast}{the center of the last class, if one wants to specifie
something different from the average of the lower and the upper
bonds}

\item{inflate}{in the case where the upper bond is infinite and
\code{xlast} is not provided, the upper bond of the last class is set
to the lower bond of the last class and the range of the
previous class times this coefficient (which default value is
one)}

\item{...}{further arguments}
}
\value{
a numerical vector
}
\description{
Convert a string (or factor) which represents a class to a value of
the underlying variable
}
\examples{

# salaire is a class of wage in the Salaires data set ; first
# extract unique values
sals <- Salaires \%>\% pull(salaire) \%>\% levels
# compute the lower bonds
sals \%>\% cls2val(0)
# lower bonds with a user specified center value for the first class
sals \%>\% cls2val(0, xfirst = 0.18) \%>\% head
# compute the upper bonds
sals \%>\% cls2val(1)
# note that the Inf upper bond is replaced by 50 + (50 - 40), ie
# the lower bond plus the range of the previous class
sals \%>\% cls2val(1, xlast = 100) \%>\% tail
# xlast is provided (the center of the last class) and the upper
# bond is adapted accordingly, which means 50 + (100 - 50) * 2 =
# 150
sals \%>\% cls2val(1, inflate = 3) \%>\% tail
# inflate is provided, so that the range of the last class is three
# times the range of the previous one
}
\author{
Yves Croissant
}
