% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rgp}
\alias{rgp}
\title{Extract of the French census}
\format{
a tibble containing :
\itemize{
\item cars : number of cars,
\item rooms : number of rooms of the housing,
\item children : number of children,
\item type : type of household ; \code{couple} or \code{monop} (for mono-parental families),
}
}
\source{
\href{https://www.insee.fr/fr/statistiques/4507685?sommaire=4508161}{INSEE}'s
website.
}
\description{
This extract of the French census gives information about a sample
of French households.
}
\keyword{datasets}
